refactor: extract mongo server and client code to separate file
